+++ model/scratch/dffml_model_scratch/anomalydetection.py
@@ -109,6 +109,81 @@ class AnomalyModelConfig:
 
 @entrypoint("anomalydetection")
 class AnomalyModel(SimpleModel):
+
+    """
+    Model for Anomaly Detection using multivariate Gaussian distribution to predict probabilities of all records in the dataset 
+    and identify outliers. F1 score is used as the evaluation metric for this model. This model works well as it recognises dependencies 
+    across various features, and works particularly well if the features have a Gaussian Distribution.
+
+    Examples
+    --------
+
+    Command line usage
+
+    Create training and test datasets
+
+    .. literalinclude:: /../model/scratch/examples/anomalydetection_ex/trainex.sh
+    
+    .. literalinclude:: /../model/scratch/examples/anomalydetection_ex/testex.sh
+
+    
+    Train the model
+
+    .. code-block:: console
+        :test:
+    
+    
+        $ dffml train \
+            - sources f=csv \
+            - source-filename trainex.csv \
+            - model anomalydetection:AnomalyModel \
+            - model-feature A:float:2 \
+            - model-predict Y:int:1  \
+            - model-directory tempdir    
+
+    
+    Assess the accuracy
+
+    .. code-block:: console
+        :test:
+
+        $ dffml accuracy \
+            -sources f=csv \
+            -source-filename testex.csv \
+            -model anomalydetection:AnomalyModel \
+            -model-feature A:float:2 \
+            -model-predict Y:int:1 \
+            -model-directory tempdir 
+    
+    
+    Make predictions
+
+    .. code-block:: console
+        :test:
+        
+        $ dffml predict all \
+            -sources f=csv \
+            -source-filename testex.csv \
+            -model anomalydetection:AnomalyModel \
+            -model-feature A:float:2 \
+            -model-predict Y:int:1 \
+            -model-directory tempdir
+
+
+    Python usage
+    
+    .. literalinclude:: /../model/scratch/examples/anomalydetection_ex/detectoutliers.py
+    
+    
+    Output
+    
+    .. code-block::
+
+        Test set F1 score : 0.8
+        Training set F1 score : 0.888888888888889
+    
+    """
+
     # The configuration class needs to be set as the CONFIG property
     CONFIG: Type = AnomalyModelConfig